rename github-electron to electron
