+++ electron/.editorconfig
